modal is component also hearts done
