+++ src/components/TrackingList/TrackingListItem.vue
@@ -7,9 +7,11 @@
       <h1 class="row-span-1 text-3xl font-bold flex items-end">
         {{ props.villager.name.charAt(0).toUpperCase() + props.villager.name.slice(1) }}
       </h1>
-      <div class="row-span-1 flex flex-row items-center">
-        <img class="object-contain h-9" src="@/assets/heart-icon.webp" />
-        <h1 class="text-4xl flex items-center ml-1 pb-1">{{ props.villager.friendshipPoints }}</h1>
+      <div class="row-span-1 flex items-center">
+        <img class="object-contain mt-1" src="@/assets/heart-icon.webp" />
+        <h1 class="text-3xl font-bold ml-1">
+          {{ isNaN(Math.floor(props.villager.friendshipPoints / 250)) ? 0 : Math.floor(props.villager.friendshipPoints / 250) }}
+        </h1>
         <img class="object-contain h-9 ml-3" v-if="villager.isMarried" src="@/assets/mermaids-pendant.png" />
       </div>
     </div>
